@@ -1,60 +1,33 @@
 {
-	"name": "alnv/prosearch",
-	"description": "ProSearch - Backend search for Contao CMS",
-	"keywords": [ "contao", "backend", "search", "filter", "tags" ],
-	"type": "contao-module",
-	"homepage": "http://backend-suche-contao.alexandernaumov.de",
-	"license": "MIT",
-	"authors": [
-		{
-			"name": "Alexander Naumov",
-			"homepage": "https://www.alexandernaumov.de",
-			"role": "Developer"
-		}
-	],
-	"require": {
-		"php": ">=5.4",
-		"contao/core-bundle": "~3.5 || ~4.1",
-		"contao-community-alliance/composer-plugin": "~2.4 || ~3.0"
-	},
-	"require-dev": {
-		"contao/easy-coding-standard": "^3.0",
-		"phpunit/phpunit": "^8.5 || ^9.5",
-		"symfony/phpunit-bridge": "^5.3 || ^6.0"
-	},
-	"support": {
-		"email": "me@alexandernaumov.de",
-		"issues": "https://github.com/alnv/prosearch/issues",
-		"wiki": "https://backend-suche-contao.alexandernaumov.de/fuer-redakteure.html"
-	},
-	"scripts": {
-		"all": [
-			"@cs-fixer",
-			"@phpunit"
-		],
-		"cs-fixer": "@php vendor/bin/ecs check src/ --fix --ansi",
-		"phpunit": "@php vendor\\bin\\phpunit --configuration phpunit.xml --coverage-text --log-junit report.xml"
-	},
-	"autoload": {
-		"classmap": [
-			"src/"
-		]
-	},
-	"replace": {
-		"contao-legacy/prosearch": "self.version"
-	},
-	"extra": {
-		"contao": {
-			"sources": {
-				"": "system/modules/prosearch"
-			}
-		}
-	},
-	"config": {
-		"allow-plugins": {
-			"contao-components/installer": true,
-			"contao-community-alliance/composer-plugin": true,
-			"dealerdirect/phpcodesniffer-composer-installer": true
-		}
-	}
+  "name": "alnv/prosearch",
+  "description": "ProSearch - Backend search for Contao CMS",
+  "keywords": ["contao", "backend", "search", "filter", "tags"],
+  "type": "contao-module",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Alexander Naumov",
+      "homepage": "https://www.alexandernaumov.de",
+      "role": "Developer"
+    }
+  ],
+  "require": {
+    "php": "^8.1",
+    "contao/core-bundle": "~4.13 || ~5.0"
+  },
+  "support": {
+    "email": "me@alexandernaumov.de",
+    "issues": "https://github.com/alnv/prosearch/issues"
+  },
+  "autoload": {
+    "psr-4": {
+      "Alnv\\ProSearchBundle\\": "src/"
+    }
+  },
+  "config": {
+    "preferred-install": "dist"
+  },
+  "extra": {
+    "contao-manager-plugin": "Alnv\\ProSearchBundle\\ContaoManager\\Plugin"
+  }
 }
