@@ -1,19 +1,21 @@
 <?php
 
-namespace ProSearch;
+namespace Alnv\ProSearchBundle\Classes;
 
 use Contao\Config;
+use Contao\Controller;
+use Contao\Database;
 use Contao\Image;
 use Contao\Input;
-use Contao\Controller;
+use Contao\StringUtil;
 
 class ProSearch extends ProSearchDataContainer
 {
-    public $coreModules = array();
-    public $notCoreModules = array('faq', 'newsletter', 'comments', 'calendar', 'news');
-    public $deletedIndexData = array();
-    public $modules = array(
+    public array $coreModules = [];
+    public array $notCoreModules = ['faq', 'newsletter', 'comments', 'calendar', 'news'];
+    public array $deletedIndexData = [];
 
+    public array $modules = array(
         'article' => array(
             'shortcut' => 'ar',
             'icon' => 'article.gif',
@@ -21,7 +23,6 @@ class ProSearch extends ProSearchDataContainer
             'searchIn' => array('title'),
             'title' => array('title')
         ),
-
         'page' => array(
             'shortcut' => 'pa',
             'tables' => array('tl_page'),
@@ -30,7 +31,6 @@ class ProSearch extends ProSearchDataContainer
             'prepareDataException' => array(array('PrepareDataException', 'prepareDataExceptions')),
             'setCustomIcon' => array(array('PrepareDataException', 'setCustomIcon')),
         ),
-
         'form' => array(
             'shortcut' => 'fo',
             'icon' => 'form.gif',
@@ -38,7 +38,6 @@ class ProSearch extends ProSearchDataContainer
             'searchIn' => array('title', 'name', 'id'),
             'title' => array('title', 'name', 'id', 'type'),
         ),
-
         'member' => array(
             'shortcut' => 'me',
             'icon' => 'member.gif',
@@ -46,7 +45,6 @@ class ProSearch extends ProSearchDataContainer
             'searchIn' => array('firstname', 'lastname', 'username', 'email', 'company', 'city', 'postal', 'gender', 'dateOfBirth'),
             'title' => array('username', 'email'),
         ),
-
         'mgroup' => array(
             'shortcut' => 'mg',
             'icon' => 'mgroup.gif',
@@ -54,7 +52,6 @@ class ProSearch extends ProSearchDataContainer
             'searchIn' => array('name'),
             'title' => array('name'),
         ),
-
         'user' => array(
             'shortcut' => 'us',
             'icon' => 'user.gif',
@@ -62,7 +59,6 @@ class ProSearch extends ProSearchDataContainer
             'searchIn' => array('username', 'name', 'email'),
             'title' => array('username'),
         ),
-
         'group' => array(
             'shortcut' => 'ug',
             'icon' => 'group.gif',
@@ -70,7 +66,6 @@ class ProSearch extends ProSearchDataContainer
             'searchIn' => array('name'),
             'title' => array('name'),
         ),
-
         'news' => array(
             'shortcut' => 'ne',
             'icon' => 'news.gif',
@@ -78,7 +73,6 @@ class ProSearch extends ProSearchDataContainer
             'searchIn' => array('title', 'subheadline', 'teaser'),
             'title' => array('headline', 'title'),
         ),
-
         'calendar' => array(
             'shortcut' => 'ev',
             'icon' => 'event.gif',
@@ -86,7 +80,6 @@ class ProSearch extends ProSearchDataContainer
             'searchIn' => array('title', 'teaser'),
             'title' => array('title'),
         ),
-
         'files' => array(
             'tables' => array('tl_files'),
             'searchIn' => array('name', 'meta'),
@@ -96,7 +89,6 @@ class ProSearch extends ProSearchDataContainer
             'setCustomShortcut' => array(array('PrepareDataException', 'setCustomShortcut')),
             'setCustomTitle' => array(array('PrepareDataException', 'setCustomTitle')),
         ),
-
         'comments' => array(
             'shortcut' => 'co',
             'icon' => 'comment.gif',
@@ -104,7 +96,6 @@ class ProSearch extends ProSearchDataContainer
             'searchIn' => array('name', 'comment'),
             'title' => array('name'),
         ),
-
         'newsletter' => array(
             'tables' => array('tl_newsletter', 'tl_newsletter_recipients'),
             'searchIn' => array('subject', 'email'),
@@ -112,7 +103,6 @@ class ProSearch extends ProSearchDataContainer
             'setCustomShortcut' => array(array('PrepareDataException', 'setCustomShortcut')),
             'setCustomIcon' => array(array('PrepareDataException', 'setCustomIcon')),
         ),
-
         'faq' => array(
             'shortcut' => 'fq',
             'icon' => 'faq.gif',
@@ -120,7 +110,6 @@ class ProSearch extends ProSearchDataContainer
             'searchIn' => array('question', 'title', 'headline'),
             'title' => array('question', 'title'),
         ),
-
         'ps_content' => array(
             'shortcut' => 'ce',
             'icon' => 'alias.gif',
@@ -131,7 +120,6 @@ class ProSearch extends ProSearchDataContainer
             'prepareDataException' => array(array('PrepareDataException', 'prepareDataExceptions')),
             'setCustomTitle' => array(array('PrepareDataException', 'setCustomTitle')),
         ),
-
         'themes' => array(
             'tables' => array('tl_module', 'tl_layout', 'tl_style_sheet', 'tl_style', 'tl_image_size'),
             'searchIn' => array('name', 'type', 'selector'),
@@ -139,9 +127,7 @@ class ProSearch extends ProSearchDataContainer
             'setCustomShortcut' => array(array('PrepareDataException', 'setCustomShortcut')),
             'setCustomIcon' => array(array('PrepareDataException', 'setCustomIcon')),
         )
-
     );
-
 
     /**
      *
@@ -151,16 +137,15 @@ class ProSearch extends ProSearchDataContainer
 
         parent::__construct();
         $this->initProSearch();
-
     }
 
     /**
      * load tables
      */
-    public function initProSearch()
+    public function initProSearch(): void
     {
         // set Vendor Modules
-        if ($GLOBALS['PS_SEARCHABLE_MODULES'] && is_array($GLOBALS['PS_SEARCHABLE_MODULES']) && !empty($GLOBALS['PS_SEARCHABLE_MODULES'])) {
+        if ($this->checkPermission($GLOBALS['PS_SEARCHABLE_MODULES']) && is_array($GLOBALS['PS_SEARCHABLE_MODULES']) && !empty($GLOBALS['PS_SEARCHABLE_MODULES'])) {
             foreach ($GLOBALS['PS_SEARCHABLE_MODULES'] as $modname => $module) {
                 if (is_array($module) && !empty($module)) {
                     $this->modules[$modname] = $module;
@@ -168,87 +153,30 @@ class ProSearch extends ProSearchDataContainer
             }
         }
 
-
         // set global
         $GLOBALS['PS_SEARCHABLE_MODULES'] = $this->modules;
-
-        // set f modules if fmodule installed
-        $this->setFModules();
 
         // set core tables
         $this->setCoreModules();
     }
 
     /**
-     *
-     */
-    public function setFModules()
-    {
-        if ($GLOBALS['PS_SEARCHABLE_MODULES'] && $GLOBALS['PS_SEARCHABLE_MODULES']['fmodule'] && is_array($GLOBALS['PS_SEARCHABLE_MODULES']['fmodule'])) {
-
-            $modulesDB = $this->Database->prepare('SELECT * FROM tl_fmodules')->execute();
-
-            while ($modulesDB->next()) {
-
-                if (!$modulesDB->tablename) continue;
-
-                $strTableCount = strlen($modulesDB->tablename);
-                $doTable = substr($modulesDB->tablename, 3, $strTableCount);
-                $wrapperTable = $modulesDB->tablename;
-                $dataTable = $modulesDB->tablename . '_data';
-
-                $icon = 'files/fmodule/assets/' . $wrapperTable . '_icon.png';
-
-                if (!file_exists(TL_ROOT . '/' . $icon)) {
-                    $this->notCoreModules[] = $doTable;
-                    $icon = 'fmodule.png';
-
-                }
-
-                $shortcut = 'fm';
-
-                if ($modulesDB->ps_shortcut) {
-                    $shortcut = $modulesDB->ps_shortcut;
-                    $GLOBALS['TL_LANG']['tl_prosearch_data']['shortcut'][$modulesDB->ps_shortcut] = $modulesDB->name;
-                }
-
-                $this->modules[$doTable] = array(
-                    'shortcut' => $shortcut,
-                    'icon' => $icon,
-                    'tables' => array($wrapperTable, $dataTable),
-                    'searchIn' => array('title', 'info', 'description'),
-                    'title' => array('title'),
-                );
-            }
-
-            $GLOBALS['PS_SEARCHABLE_MODULES'] = $this->modules;
-
-        }
-    }
-
-    /**
      * @return array
      * load all searchable modules
      */
-    public function loadModules()
+    public function loadModules(): array
     {
 
         $this->initProSearch();
 
-        $return = array();
-
-        // set core modules
+        $return = [];
         $coreModules = $this->coreModules;
 
-
-        // push dca' into $searchDataContainerArr
         foreach ($coreModules as $tablename => $coreModule) {
 
-            // load dca
             $this->loadDataContainer($tablename);
 
-            // break up if dca not exist
-            if (!$GLOBALS['TL_DCA'][$tablename]) {
+            if (!($GLOBALS['TL_DCA'][$tablename] ?? '')) {
                 continue;
             }
 
@@ -259,33 +187,26 @@ class ProSearch extends ProSearchDataContainer
         return $return;
     }
 
-
-    /**
-     *
-     */
-    public function setCoreModules()
+    public function setCoreModules(): void
     {
 
         foreach ($this->modules as $k => $module) {
 
-            $label = $GLOBALS['TL_LANG']['MOD'][$k];
+            $label = ($GLOBALS['TL_LANG']['MOD'][$k] ?? '');
             $this->modules[$k]['label'] = $label;
 
             foreach ($module['tables'] as $table) {
-                $this->coreModules[$table] = $label[0] ? $label[0] : '[no-label-found]';
+                $this->coreModules[$table] = $label[0] ?? '[no-label-found]';
             }
         }
     }
 
-    /**
-     *
-     */
     public function deleteModulesFromIndex()
     {
 
-        $activeModules = deserialize(Config::get('searchIndexModules')) ? deserialize(Config::get('searchIndexModules')) : array();
+        $activeModules = StringUtil::deserialize(Config::get('searchIndexModules')) ? StringUtil::deserialize(Config::get('searchIndexModules')) : array();
         $coreModulesArr = Helper::pluckModules($this->coreModules);
-        $toDeleteArr = array_diff($coreModulesArr, $activeModules);
+        $toDeleteArr = \array_diff($coreModulesArr, $activeModules);
 
         $i = 0;
         $whereStr = '';
@@ -293,12 +214,9 @@ class ProSearch extends ProSearchDataContainer
         foreach ($toDeleteArr as $key => $value) {
             if ($i == 0) {
                 $whereStr = 'WHERE dca = "' . $value . '"';
-
             } else {
-
                 $whereStr .= ' OR dca = "' . $value . '"';
             }
-
             $i++;
         }
 
@@ -306,22 +224,21 @@ class ProSearch extends ProSearchDataContainer
             return null;
         }
 
-        $this->Database->prepare('DELETE FROM tl_prosearch_data ' . $whereStr . '')->execute();
-
+        Database::getInstance()->prepare('DELETE FROM tl_prosearch_data ' . $whereStr)->execute();
     }
 
     /**
      * @param $strName
      */
-    public function createOnSubmitCallback($strName)
+    public function createOnSubmitCallback($strName): void
     {
 
         $coreModulesArr = Helper::pluckModules($this->coreModules);
 
         if (in_array($strName, $coreModulesArr) && $GLOBALS['TL_DCA'][$strName]) {
-            $GLOBALS['TL_DCA'][$strName]['config']['onsubmit_callback'][] = array('ProSearch', 'sendDataToIndex');
-            $GLOBALS['TL_DCA'][$strName]['config']['oncut_callback'][] = array('ProSearch', 'sendDataToIndex');
-            $GLOBALS['TL_DCA'][$strName]['config']['ondelete_callback'][] = array('ProSearch', 'deleteDataFromIndex');
+            $GLOBALS['TL_DCA'][$strName]['config']['onsubmit_callback'][] = array(ProSearch::class, 'sendDataToIndex');
+            $GLOBALS['TL_DCA'][$strName]['config']['oncut_callback'][] = array(ProSearch::class, 'sendDataToIndex');
+            $GLOBALS['TL_DCA'][$strName]['config']['ondelete_callback'][] = array(ProSearch::class, 'deleteDataFromIndex');
         }
     }
 
@@ -340,35 +257,30 @@ class ProSearch extends ProSearchDataContainer
         $colname = 'id';
 
         // get act
-        $act = Input::get('act');
+        $act = Input::get('act') ?: '';
 
         // if cut
-        if ($act && $act == 'cut') {
+        if ($act == 'cut') {
 
             $id = Input::get('id');
 
-            if(Input::get('do') == 'files')
-            {
+            if (Input::get('do') == 'files') {
                 $tablename = 'tl_files';
                 $id = $value;
                 $colname = 'path';
             }
 
-            if(!$tablename)
-            {
+            if (!$tablename) {
                 return;
             }
 
-            $dcaArr = $this->Database->prepare('SELECT * FROM ' . $tablename . ' WHERE '.$colname.' = ?')->execute($id)->row();
-
+            $dcaArr = Database::getInstance()->prepare('SELECT * FROM ' . $tablename . ' WHERE ' . $colname . ' = ?')->execute($id)->row();
         }
 
-        //
         $arr = array();
+        $data = $this->prepareIndexData($dcaArr, ($GLOBALS['TL_DCA'][$tablename] ?? []), $tablename);
 
-        $data = $this->prepareIndexData($dcaArr, $GLOBALS['TL_DCA'][$tablename], $tablename);
-
-        if ($data == false) {
+        if ($data === false) {
             return;
         }
 
@@ -383,23 +295,18 @@ class ProSearch extends ProSearchDataContainer
      * @param $doTable
      * @return string
      */
-    public function getIcon($doTable)
+    public function getIcon($doTable): string
     {
+
         $icon = '';
+        $path = 'bundles/alnvprosearch/images/';
 
-        $path = 'system/modules/prosearch/assets/images/';
 
-        if ((version_compare(VERSION, '4.0', '>=') && !$GLOBALS['PS_NO_COMPOSER'] && $GLOBALS['PS_NO_COMPOSER'] != true)) {
-            $path = 'bundles/prosearch/images/';
-        }
-
-        if (!in_array($doTable, $this->notCoreModules))
-        {
+        if (!in_array($doTable, $this->notCoreModules)) {
             $path = '';
         }
 
-        if ($this->modules[$doTable] && $this->modules[$doTable]['icon'])
-        {
+        if ($this->modules[$doTable] && $this->modules[$doTable]['icon']) {
             $icon = Image::getHtml($path . $this->modules[$doTable]['icon']);
         }
 
@@ -429,14 +336,14 @@ class ProSearch extends ProSearchDataContainer
             'dca' => $table,
             'tstamp' => time(),
             'doTable' => $doTable,
-            'ptable' => $dca['config']['ptable'] ? $dca['config']['ptable'] : '',
+            'ptable' => $dca['config']['ptable'] ?: '',
             'ctable' => $dca['config']['ctable'] ? serialize($dca['config']['ctable']) : '',
             'docId' => $db['id'],
-            'pid' => isset($db['pid']) ? $db['pid'] : 0,
-            'extension' => $db['extension'] ? $db['extension'] : '',
-            'tags' => $db['ps_tags'] ? $db['ps_tags'] : '',
-            'blocked' => $db['ps_block_item'] ? $db['ps_block_item'] : '',
-            'blocked_ug' => $db['ps_block_usergroup'] ? $db['ps_block_usergroup'] : ''
+            'pid' => $db['pid'] ?? 0,
+            'extension' => $db['extension'] ?: '',
+            'tags' => $db['ps_tags'] ?: '',
+            'blocked' => $db['ps_block_item'] ?: '',
+            'blocked_ug' => $db['ps_block_usergroup'] ?: ''
         );
 
 
@@ -445,7 +352,7 @@ class ProSearch extends ProSearchDataContainer
          */
         $shortcut = '';
         if ($this->modules[$doTable]) {
-            $shortcut = $this->modules[$doTable]['shortcut'] ? $this->modules[$doTable]['shortcut'] : '';
+            $shortcut = $this->modules[$doTable]['shortcut'] ?: '';
         }
         $arr['shortcut'] = $shortcut;
         if ($this->modules[$doTable] && is_array($this->modules[$doTable]['setCustomShortcut'])) {
@@ -462,10 +369,8 @@ class ProSearch extends ProSearchDataContainer
         /**
          * set custom icon callbacks
          */
-        if ($this->modules[$doTable] && is_array($this->modules[$doTable]['setCustomIcon']))
-        {
-            foreach ($this->modules[$doTable]['setCustomIcon'] as $callable)
-            {
+        if ($this->modules[$doTable] && is_array($this->modules[$doTable]['setCustomIcon'])) {
+            foreach ($this->modules[$doTable]['setCustomIcon'] as $callable) {
                 $this->import($callable[0]);
                 $this->modules[$doTable]['icon'] = $this->{$callable[0]}->{$callable[1]}($table, $db, $arr, $dca);
 
@@ -553,17 +458,14 @@ class ProSearch extends ProSearchDataContainer
 
         $colsSearchContent = $this->modules[$doTable]['searchIn'];
 
-        array_unshift($colsSearchContent, 'ps_search_content');
+        \array_unshift($colsSearchContent, 'ps_search_content');
         $colsSearchContent = $colsSearchContent ? $colsSearchContent : array();
 
         // addDescriptionToSearchContent
         $textCols = array('text', 'description');
-        if (Config::get('addDescriptionToSearchContent'))
-        {
-            foreach($textCols as $col)
-            {
-                if(in_array($col, $db))
-                {
+        if (Config::get('addDescriptionToSearchContent')) {
+            foreach ($textCols as $col) {
+                if (in_array($col, $db)) {
                     $colsSearchContent[] = $col;
                 }
             }
@@ -573,12 +475,11 @@ class ProSearch extends ProSearchDataContainer
 
         foreach ($colsSearchContent as $content) {
 
-            if(!$db[$content])
-            {
+            if (!$db[$content]) {
                 continue;
             }
 
-            $ct = deserialize($db[$content]);
+            $ct = StringUtil::deserialize($db[$content]);
 
             if (is_array($ct) && !empty($ct)) {
                 $meta = Helper::parseStrForMeta($db[$content]);
@@ -587,9 +488,7 @@ class ProSearch extends ProSearchDataContainer
 
             }
 
-
             if ($db[$content] && (is_string($db[$content]) || is_numeric($db[$content]))) {
-
                 $strContent .= ' ' . $db[$content];
             }
         }
@@ -597,18 +496,11 @@ class ProSearch extends ProSearchDataContainer
         $strContent = Controller::replaceInsertTags($strContent);
         $strContent = strip_tags($strContent);
         $strContent = trim($strContent);
-        $strContent = mb_strtolower($strContent);
 
-        return $strContent;
+        return mb_strtolower($strContent);
 
     }
 
-    /**
-     * @param $db
-     * @param $doTable
-     * @param $table
-     * @return null|string
-     */
     public function setTitle($db, $doTable, $table)
     {
         // sorted by priority
@@ -618,7 +510,7 @@ class ProSearch extends ProSearchDataContainer
 
         $colsForTitle = $this->modules[$doTable]['title'];
         array_unshift($colsForTitle, 'ps_title');
-        $colsForTitle = $colsForTitle ? $colsForTitle : array();
+        $colsForTitle = $colsForTitle ?: array();
 
         // hook for custom title
         if ($this->modules[$doTable] && is_array($this->modules[$doTable]['setCustomTitle'])) {
@@ -635,12 +527,11 @@ class ProSearch extends ProSearchDataContainer
 
         foreach ($colsForTitle as $title) {
 
-            if(!$db[$title])
-            {
+            if (!$db[$title]) {
                 continue;
             }
-            
-            $ct = deserialize($db[$title]);
+
+            $ct = StringUtil::deserialize($db[$title]);
 
             // check if value is serialize
             if (is_array($ct) && !empty($ct)) {
@@ -693,24 +584,16 @@ class ProSearch extends ProSearchDataContainer
         }
     }
 
-    /**
-     * @param $dca
-     * @return null
-     */
-    public function clearSearchIndexTable($dca)
+    public function clearSearchIndexTable($dca): void
     {
         if (!$dca) {
-            return null;
+            return;
         }
 
         $this->Database->prepare('DELETE FROM tl_prosearch_data WHERE dca = ?')->execute($dca);
     }
 
-    /**
-     * @param $data
-     * @param $dca
-     */
-    public function saveSingleIndexIntoDB($data, $dca)
+    public function saveSingleIndexIntoDB($data, $dca): void
     {
 
         foreach ($data as $arr) {
@@ -732,10 +615,10 @@ class ProSearch extends ProSearchDataContainer
             $this->Database->prepare('INSERT INTO tl_prosearch_data(' . $cols . ') VALUES (' . $placeholder . ')')->execute($values);
 
         }
-
     }
 
-    public function deleteDataFromIndex($dc) {
+    public function deleteDataFromIndex($dc): void
+    {
 
         $tablename = $dc->table;
         $docId = $dc->activeRecord->id;
@@ -761,14 +644,14 @@ class ProSearch extends ProSearchDataContainer
             $this->deletedIndexData[] = $pDataDB->row();
         }
 
-
         //delete all childs
         $this->deleteChildrenFromIndex($tablename, $docId);
         $this->clearIndex();
 
     }
 
-    public function deleteChildrenFromIndex($ptable, $pid) {
+    public function deleteChildrenFromIndex($ptable, $pid)
+    {
 
         $cDataDB = $this->Database->prepare('SELECT * FROM tl_prosearch_data WHERE ptable = ? AND pid = ?')->execute($ptable, $pid);
 
@@ -779,7 +662,7 @@ class ProSearch extends ProSearchDataContainer
                 $id = $cDataDB->docId;
                 $activeModules = deserialize(Config::get('searchIndexModules'));
                 $ctables = deserialize($cDataDB->ctable);
-                
+
                 if (in_array($table, $activeModules)) {
                     $this->deleteChildrenFromIndex($table, $id);
                 }
@@ -794,37 +677,34 @@ class ProSearch extends ProSearchDataContainer
                 }
 
                 $this->deletedIndexData[] = $cDataDB->row();
-
             }
-
         }
     }
 
-    public function clearIndex() {
+    public function clearIndex()
+    {
 
         foreach ($this->deletedIndexData as $indexData) {
-
             $this->Database->prepare('DELETE FROM tl_prosearch_data WHERE id = ?')->execute($indexData['id']);
         }
     }
 
-    public function ajaxRequest() {
+    public function ajaxRequest()
+    {
 
         if (Input::get('ajaxRequestForProSearch') && Input::get('ajaxRequestForProSearch') == 'getSearchIndex') {
 
             $q = Input::get('searchQuery');
 
             $header = array(
-
-                'q' => $q ? $q : ''
+                'q' => $q ?: ''
             );
 
             $results = array();
             $results['response'] = $this->getSearchDataFromIndex($header);
 
-            $results['shortcut_labels'] = array();
             $this->loadLanguageFile('tl_prosearch_data');
-            $results['shortcut_labels'] = $GLOBALS['TL_LANG']['tl_prosearch_data']['shortcut'];
+            $results['shortcut_labels'] = $GLOBALS['TL_LANG']['tl_prosearch_data']['shortcut'] ?? '';
 
             header('Content-type: application/json');
             echo json_encode($results, 512);
@@ -987,21 +867,22 @@ class ProSearch extends ProSearchDataContainer
     private function checkPermission($searchItem)
     {
 
-        $group = $this->User->groups ? $this->User->groups : array();
+
+        $group = $this->User->groups ?? [];
 
         if (empty($group)) {
             return true;
         }
 
         if (!$this->User->isAdmin) {
-            $blocked_ug = $searchItem['blocked_ug'] ? $searchItem['blocked_ug'] : array();
-            $blocked_ug = deserialize($blocked_ug);
+            $blocked_ug = $searchItem['blocked_ug'] ?? [];
+            $blocked_ug = StringUtil::deserialize($blocked_ug, true);
 
             if (empty($blocked_ug)) {
                 return true;
             }
 
-            if (array_intersect($group, $blocked_ug) > 0) {
+            if (\array_intersect($group, $blocked_ug) > 0) {
                 return false;
             }
 
